hackthon-ragllm/lambda/requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
python-dotenv
boto3